implemented login logout feature, logged in user will only see logout button, logged out user will only see login button in nav bar
